discoveryserver: rename from discovery
Make this more clear what it is.
