+++ discoveryserver/.gitignore
