+++ discoveryserver/handlers/home.go
