+++ discoveryserver/handlers/httperror/httperror.go
