+++ discoveryserver/handlers/robots.go
