+++ discoveryserver/handlers/token.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"go.etcd.io/etcd/discovery/handlers/httperror"
+	"go.etcd.io/etcd/discoveryserver/handlers/httperror"
 	"go.etcd.io/etcd/etcdserver/api/v2store"
 
 	"github.com/prometheus/client_golang/prometheus"
