+++ discoveryserver/integration/server.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	discoveryhttp "go.etcd.io/discovery.etcd.io/http"
+	discoveryhttp "go.etcd.io/etcd/discoveryserver/http"
 
 	"go.etcd.io/etcd/embed"
 	"go.etcd.io/etcd/etcdserver/api/v3client"
