+++ discoveryserver/tests.sh
